requirements.txt
Pillow
accelerate==1.4.0
av==14.0.0
bitsandbytes
botocore
diffusers==0.31.0
einops
fsspec
numpy
pandas
protobuf==4.21.6
pyarrow
safetensors
sentencepiece
scipy
tensorboard
torch-fidelity
torchmetrics==1.6.0
tqdm
transformers==4.48.0
transforms3d
Ninja
chamferdist
timm
rotary-embedding-torch
opencv-python
wandb
edict
